Added basic database connection files.
	new file:   db_config.php
Contains the database parameters required in the connection string for connection to the database. Defines the database name, user, password, server and server port.

	new file:   db_connect.php
Contains code to connect to and disconnect from the database. A class is defined which contains methods to connect to and disconnect from the database using the database parameters defined in db_config.php.